Revert "Make kunzip extract a zipfile with incomplete local file header"
This reverts commit d7213b86386c5e4d8524801d424bd22e10bfeebb.

After reading the ZIP file specs at
http://www.pkware.com/business_and_developers/developer/popups/appnote.txt
it is clear that the OpenOffice files are absolutely correct.  OpenOffice
seems to write its documents in a stream and does not put the files' crc32
and sizes in to the local file header, but into a data descriptor that
follows the file contents.  This is correctly indicated by the general
purpose bit flag in the local file header.